+++ creme/compose/blacklist.py
@@ -8,7 +8,7 @@ class Blacklister(base.Transformer):
     """Subsets a set of features by applying a blacklist.
 
     Parameters:
-        blacklist (str or list): Key(s) to discard.
+        blacklist (strs): Key(s) to discard.
 
     Example:
 
@@ -17,7 +17,7 @@ class Blacklister(base.Transformer):
             >>> from creme import compose
 
             >>> x = {'a': 42, 'b': 12}
-            >>> compose.Blacklister(['a', 'zoidberg']).transform_one(x)
+            >>> compose.Blacklister('a', 'zoidberg').transform_one(x)
             {'b': 12}
 
             >>> compose.Blacklister('b').transform_one(x)
@@ -25,13 +25,11 @@ class Blacklister(base.Transformer):
 
     """
 
-    def __init__(self, blacklist=None):
-        if blacklist is None:
-            blacklist = []
-        self.blacklist = set(blacklist if isinstance(blacklist, (list, tuple)) else [blacklist])
+    def __init__(self, *blacklist):
+        self.blacklist = set(blacklist)
 
     def transform_one(self, x):
         return {i: x[i] for i in set(x.keys()) - self.blacklist}
 
     def __str__(self):
-        return self.__class__.__name__ + f'({list(self.blacklist)})'
+        return '~' + str(sorted(self.blacklist))