+++ creme/compose/whitelist.py
@@ -8,7 +8,7 @@ class Whitelister(base.Transformer):
     """Subsets a set of features by applying a whitelist.
 
     Parameters:
-        whitelist (str or list): Key(s) to keep.
+        whitelist (strs): Key(s) to keep.
 
     Example:
 
@@ -17,7 +17,7 @@ class Whitelister(base.Transformer):
             >>> from creme import compose
 
             >>> x = {'a': 42, 'b': 12}
-            >>> compose.Whitelister(['a', 'zoidberg']).transform_one(x)
+            >>> compose.Whitelister('a', 'zoidberg').transform_one(x)
             {'a': 42}
 
             >>> compose.Whitelister('b').transform_one(x)
@@ -25,13 +25,11 @@ class Whitelister(base.Transformer):
 
     """
 
-    def __init__(self, whitelist=None):
-        if whitelist is None:
-            whitelist = []
-        self.whitelist = set(whitelist if isinstance(whitelist, (list, tuple)) else [whitelist])
+    def __init__(self, *whitelist):
+        self.whitelist = set(whitelist)
 
     def transform_one(self, x):
         return {i: x[i] for i in set(x.keys()) & self.whitelist}
 
     def __str__(self):
-        return self.__class__.__name__ + f'({list(self.whitelist)})'
+        return str(sorted(self.whitelist))