+++ creme/optim/ada_bound.py
@@ -3,17 +3,13 @@ import math
 
 from .. import utils
 from . import base
-# from . import lr_schedule
 
 
 __all__ = ['AdaBound']
 
 
 class AdaBound(base.Optimizer):
-    """AdaBound Optimizer.
-
-    Adaptive Gradient Methods with Dynamic Bound of Learning Rate:
-            https://openreview.net/forum?id=Bkg3g2R9FX
+    """AdaBound optimizer.
 
     Example:
 
@@ -41,6 +37,10 @@ class AdaBound(base.Optimizer):
 
             >>> model_selection.online_score(X_y, model, metric)
             F1Score: 0.963277
+
+    References:
+        1. `Adaptive Gradient Methods with Dynamic Bound of Learning Rate <https://openreview.net/forum?id=Bkg3g2R9FX>`_
+
     """
 
     def __init__(self, lr=1e-3, beta_1=0.9, beta_2=0.999, eps=1e-8, gamma=1e-3, final_lr=0.1):