+++ creme/types.py
@@ -1,4 +1,0 @@
-import typing
-
-
-Label = typing.Union[str, int]